msm: kgsl: Use the correct virtual address for pmem regions
We need to pass in the virtual (kernel) address to the cache
flush functions for pmem regions.  While the code we have now
is functionally correct, we are getting to the right answer
the wrong way and this breaks when the PMEM regions are moved out
of the oversight of the kernel memory manager.  Properly get the
virtual address to future proof us against changes.

Change-Id: I0aec5912c4d25f56809e1e2a4869aaf4a99fe2be
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>