msm: kgsl: unravel the 2d register functions
Because not all of the 2d registers are directly accessible,
the functions for accessing them had gotten a little bit
confusing. This removes the apparent mutual recursion at the
expense of a little more code.

Also, the 2 step registers are now guarded by a spinlock to
prevent a possible race.

Change-Id: I833b896a8dfa2a85079899d4f1b057be91078ab4
Signed-off-by: Jeremy Gebben <jgebben@codeaurora.org>