msm: kgsl: Implement static + dynamic pagetable pool scheme
We have tried a pagetable pool implementation with a strict
number of entries, and we have tried one with a dynamic number
of entries. The former required too much memory and was too
rigid, and the latter fell down on busy and memory constrained
systems. So this should be the best of both worlds - allocate
a configurable pool of pagetables at init time.  This is the
minimum guaranteed number of pagetables we will support.
Additional pagetables over this number will be allocated dynamically
if there is enough memory to do so.

Change-Id: Ic7431003afdb99efbf03e30e771d68e93d4d1ccf
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>